refactor: 서비스 계층 책임 분리
