+++ src/main/java/roomescape/reservation/service/ReservationManager.java
@@ -1,8 +1,10 @@
 package roomescape.reservation.service;
 
+import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.InvalidArgumentException;
 import roomescape.member.domain.Member;
 import roomescape.member.service.MemberQueryService;
 import roomescape.reservation.domain.Reservation;
@@ -14,7 +16,7 @@ import roomescape.theme.domain.Theme;
 import roomescape.theme.service.ThemeQueryService;
 import roomescape.time.service.ReservationTimeQueryService;
 
-@Service
+@Component
 @RequiredArgsConstructor
 public class ReservationManager {
 
@@ -25,14 +27,25 @@ public class ReservationManager {
 
     @Transactional
     public Reservation reserve(ReserveCommand reserveCommand) {
+        isAlreadyReservedTime(reserveCommand.date(), reserveCommand.timeId());
+        Reservation reserve = reservationFrom(reserveCommand);
+
+        return reservationRepository.save(reserve);
+    }
+
+    private Reservation reservationFrom(ReserveCommand reserveCommand) {
         Theme theme = themeQueryService.getTheme(reserveCommand.themeId());
         Member member = memberQueryService.getMember(reserveCommand.memberId());
         ReservationDateTime reservationDateTime = ReservationDateTime.create(new ReservationDate(reserveCommand.date()),
                 reservationTimeQueryService.getReservationTime(reserveCommand.timeId()));
 
-        Reservation reserve = Reservation.reserve(member, reservationDateTime, theme);
+        return Reservation.reserve(member, reservationDateTime, theme);
+    }
 
-        return reservationRepository.save(reserve);
+    private void isAlreadyReservedTime(LocalDate date, Long timeId) {
+        if (reservationRepository.existsByDateAndTimeId(date, timeId)) {
+            throw new InvalidArgumentException("이미 예약이 존재하는 시간입니다.");
+        }
     }
 
     @Transactional
